move json schea to src
