+++ src/monika-config-schema.json
@@ -42,6 +42,12 @@
             "type": "integer",
             "description": "Postgres port to connect to",
             "default": "5432"
+          },
+          "database": {
+            "title": "Database",
+            "type": "string",
+            "description": "Name of the datatabase",
+            "default": "dbname"
           }
         }
       }
