+++ tsconfig.json
@@ -5,7 +5,7 @@
     "importHelpers": true,
     "module": "commonjs",
     "outDir": "lib",
-    // "rootDir": "src",
+    "rootDir": "src",
     "strict": true,
     "target": "es2017",
     "skipLibCheck": true,
@@ -13,5 +13,5 @@
     "resolveJsonModule": true,
     "typeRoots": ["node_modules/@types", "custom_types"]
   },
-  "include": ["src/**/*", "src/**/*.json", "*.json"]
+  "include": ["src/**/*", "src/**/*.json"]
 }
